feat: comprehensive proto v1 migration and unified intelligence platform repositioning
- Complete proto v1 migration with all legacy references eliminated
- Enhanced v1 schema with missing types: QuantizationConfig, CollectionConfig, etc.
- Enabled v1 SQL filters with comprehensive operator support
- Repositioned as unified intelligence platform creating new AI solution category
- Updated README, logo, and cover for multimodal, serverless, graph+vector+SKS positioning
- Marketing optimization for lead generation targeting decision makers
- Documentation updates emphasizing unified intelligence capabilities

+++ src/storage/engines/core/formats/columnar/serialization.rs
@@ -515,7 +515,7 @@ impl ColumnarSerializer {
             .map(|(i, vector)| VectorRecord {
                 id: format!("record_{}", i), // Placeholder - would come from ID column
                 vector,
-                timestamp: chrono::Utc::now().timestamp() as u32,
+                timestamp: chrono::Utc::now().timestamp(),
                 ..Default::default()
             })
             .collect();
